Loads of code cleanup.
